re organize code and make it into a static library
