+++ tldr_app/.gitignore
@@ -1,0 +1,1 @@
+.vscode
