+++ tldr_app/tldr-dekstop/tldr_cpp/CMakeLists.txt
@@ -2,7 +2,7 @@
 cmake_minimum_required(VERSION 3.10)
 project(tldr_cpp)
 
-set(SOURCE_DIR ${CMAKE_SOURCE_DIR}/src/c++)
+set(SOURCE_DIR ${CMAKE_SOURCE_DIR}/src)
 include_directories(${SOURCE_DIR})
 
 set(CMAKE_CXX_STANDARD 20)
@@ -24,51 +24,58 @@ if (PkgConfig_FOUND)
     pkg_check_modules(POPPLER REQUIRED poppler-cpp)
 endif ()
 
-# Include directories
-include_directories()
+# Add linker flags for library paths
+link_directories(
+    /opt/homebrew/opt/libpq/lib
+    /opt/homebrew/opt/libpqxx/lib
+    /opt/homebrew/opt/poppler/lib
+)
+
+# Create static library
+add_library(lib_tldr STATIC
+    ${SOURCE_DIR}/lib_tldr/lib_tldr.cpp
+    ${SOURCE_DIR}/lib_tldr/lib_tldr.h
+    ${SOURCE_DIR}/lib_tldr/constants.h
+    ${SOURCE_DIR}/lib_tldr/db/sqlite_database.cpp
+    ${SOURCE_DIR}/lib_tldr/db/sqlite_database.h
+    ${SOURCE_DIR}/lib_tldr/db/database.h
+    ${SOURCE_DIR}/lib_tldr/db/postgres_database.cpp
+    ${SOURCE_DIR}/lib_tldr/db/postgres_database.h
+    ${SOURCE_DIR}/lib_tldr/db/connection_pool.h
+    ${SOURCE_DIR}/tldr.cpp
+    ${SOURCE_DIR}/tldr.h
+)
+
+# Include directories for the library
+target_include_directories(lib_tldr PRIVATE
+    ${PostgreSQL_INCLUDE_DIRS}
+    ${PQXX_INCLUDE_DIRS}
+    ${SQLite3_INCLUDE_DIRS}
+    /opt/homebrew/opt/libpq/include
+    /opt/homebrew/opt/libpqxx/include
+    /opt/homebrew/opt/poppler/include
+    /opt/homebrew/opt/curl/include
+    /opt/homebrew/opt/nlohmann-json/include
+)
+
+# Link libraries for the static library
+target_link_libraries(lib_tldr PRIVATE
+    ${PostgreSQL_LIBRARIES}
+    ${PQXX_LIBRARIES}
+    SQLite::SQLite3
+    SQLiteCpp
+    nlohmann_json::nlohmann_json
+    CURL::libcurl
+    pqxx
+    pq
+    poppler-cpp
+)
 
 # Add the executable
-file(GLOB_RECURSE SRC_FILES 
-    ${SOURCE_DIR}/*.cpp 
-    ${SOURCE_DIR}/*.h
-    ${SOURCE_DIR}/db/*.cpp
-    ${SOURCE_DIR}/db/*.h
-)
+add_executable(tldr_cpp ${SOURCE_DIR}/main.cpp)
 
-add_executable(tldr_cpp ${SRC_FILES}
-    src/c++/main.cpp
-    src/c++/main.h
-    src/c++/db/database.h
-    src/c++/db/sqlite_database.h
-    src/c++/db/sqlite_database.cpp
-    src/c++/db/postgres_database.h
-    src/c++/db/postgres_database.cpp
-)
+# Link the static library with the executable
+target_link_libraries(tldr_cpp PRIVATE lib_tldr)
 
-
-# Include directories
-target_include_directories(tldr_cpp PRIVATE
-        ${PostgreSQL_INCLUDE_DIRS}
-        ${PQXX_INCLUDE_DIRS}
-        ${SQLite3_INCLUDE_DIRS}
-        /opt/homebrew/opt/libpq/include
-        /opt/homebrew/opt/libpqxx/include
-        /opt/homebrew/opt/poppler/include
-        /opt/homebrew/opt/curl/include
-        /opt/homebrew/opt/nlohmann-json/include
-)
-
-target_link_options(tldr_cpp PRIVATE
-        -L/opt/homebrew/opt/libpq/lib/
-        -L/opt/homebrew/opt/libpqxx/lib/
-        -L/opt/homebrew/opt/poppler/lib/
-        -lpq
-        -lpoppler-cpp
-        -lcurl
-)
-## Link libraries
-target_link_libraries(tldr_cpp PRIVATE
-        ${PostgreSQL_LIBRARIES} ${PQXX_LIBRARIES}
-        SQLite::SQLite3
-        SQLiteCpp
-)
+# Include directories for the executable
+target_include_directories(tldr_cpp PRIVATE ${SOURCE_DIR})
